pyproject.toml
[project]
name = "local-rlm"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "aiosqlite>=0.22.1",
    "chromadb>=0.3.23",
    "ddgs>=9.10.0",
    "dspy-ai>=3.0.4",
    "fastapi>=0.128.0",
    "google-generativeai>=0.8.6",
    "ipykernel>=7.1.0",
    "ipython>=9.9.0",
    "markdown>=3.7",
    "matplotlib>=3.9.0",
    "nbconvert>=7.16.6",
    "notebook>=7.5.1",
    "openai>=2.14.0",
    "openpyxl>=3.1.5",
    "weasyprint>=63.1",
    "pandas>=2.3.3",
    "pdfplumber>=0.11.9",
    "pypdf>=6.5.0",
    "python-docx>=1.2.0",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.21",
    "seaborn>=0.13.0",
    "uvicorn[standard]>=0.40.0",
    "websockets>=15.0.1",
    "scipy>=1.16.3",
    "scikit-learn>=1.8.0",
    "statsmodels>=0.14.6",
    "torch>=2.9.1",
    "torchvision>=0.24.1",
    "torchaudio>=2.9.1",
]

[dependency-groups]
dev = [
    "playwright>=1.57.0",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-playwright>=0.7.2",
    "pytest-timeout>=2.4.0",
]

[tool.pytest.ini_options]
pythonpath = "."
testpaths = ["tests"]
filterwarnings = [
    "ignore:Pydantic serializer warnings:UserWarning",
    "ignore:'asyncio.iscoroutinefunction' is deprecated:DeprecationWarning"
]

